services/user_service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
boto3==1.34.34
requests==2.31.0
httpx==0.27.0
loguru==0.7.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pytz==2024.1
google-generativeai==0.8.3
google-genai==0.3.0
Pillow==10.2.0
firebase-admin==6.4.0

# Audio validation & quality improvement
numpy==1.26.4
soundfile==0.12.1
pydub==0.25.1
amazon-transcribe==0.6.2